Split if and if let
