+++ presentation/chapters/shared/code/syntax/14.rs
@@ -1,8 +1,14 @@
-fn do_thing() -> bool {
-    "Forgotten semicolon"
-    true
+use foo::foo;
+
+mod foo {
+    pub fn foo() {
+        // ...
+    }
 }
 
+// Will try to open `./bar.rs` relative to this file.
+pub mod bar;
+
 fn main() {
-    do_thing()
+    foo()
 }
