+++ presentation/chapters/shared/code/syntax/15.rs
@@ -1,0 +1,8 @@
+fn do_thing() -> bool {
+    "Forgotten semicolon"
+    true
+}
+
+fn main() {
+    do_thing()
+}
