+++ presentation/chapters/shared/code/syntax/4.rs
@@ -1,15 +1,7 @@
 fn main() {
-    let mut value = 0;
-    // Loop with break
-    loop {
-        if value >= 10 {
-            break;
-        }
-        value += 1;
-    }
-    // Break on conditional
-    while value <= 10 {
-        value += 1;
+    let maybe_value = Some(2);
+    
+    if let Some(value) = maybe_value {
         // ...
-    }
+    } else { /* ... */ }
 }
