+++ presentation/chapters/shared/code/syntax/9.rs
@@ -1,8 +1,10 @@
-fn with_lifetimes<'a>(thing: &'a str) -> &'a str {
-    thing
+// &mut denotes a mutable borrow
+fn accepts_borrow(thing: &mut u32) {
+    *thing += 1
 }
 
 fn main() {
-    let foo = "foo";
-    println!("{}", with_lifetimes(foo))
+    let mut value = 1;
+    accepts_borrow(&mut value);
+    println!("{}", value)
 }
